+++ lib/core/widgets/buttons/go_button.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_question/core/widgets/pressable.dart';
 
-part 'go_button/_go_button_hsl.dart';
-part 'go_button/_go_button_colors.dart';
-part 'go_button/_go_button_painter.dart';
-part 'go_button/_go_button_content.dart';
-part 'go_button/_go_button_widget.dart';
+part 'go_button/go_button_hsl.dart';
+part 'go_button/go_button_colors.dart';
+part 'go_button/go_button_painter.dart';
+part 'go_button/go_button_content.dart';
+part 'go_button/go_button_widget.dart';
