+++ lib/features/home/presentation/widgets/profile_button.dart
@@ -4,9 +4,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_question/config/theme/ui_constants.dart';
 import 'package:go_question/features/score/presentation/cubit/score_cubit.dart';
 
-part 'profile_button/_profile_avatar.dart';
-part 'profile_button/_profile_user_info.dart';
-part 'profile_button/_profile_score_badge.dart';
+part 'profile_button/profile_avatar.dart';
+part 'profile_button/profile_user_info.dart';
+part 'profile_button/profile_score_badge.dart';
 
 // ─────────────────────────────────────────────────────────────────────────────
 // ProfileButton — карточка профиля на главном экране.
