+++ lib/features/home/presentation/widgets/profile_button/profile_score_badge.dart
